[Fix] result page modal component 구분
