+++ front/components/search/result/Result.tsx
@@ -1,4 +1,5 @@
-import Modal from "@modal/Modal";
+import { useState } from "react";
+
 import {
   FOOTER_HEIGHT,
   FULL_HEIGHT,
@@ -7,7 +8,6 @@ import {
   SUB_COLOR,
 } from "@utils/constant";
 import React from "react";
-import { useState } from "react";
 import {
   Bottom,
   SmallContainer as Caution,
@@ -54,6 +54,7 @@ import {
   CloseModalBtn,
   ModalContentScroll,
 } from "./Result.style";
+import ResultModal from "./ResultModal";
 
 const tempData: { [key in string]: string } = {
   title: "한미아스피린장용정 100밀리그램",
@@ -71,15 +72,13 @@ const tempData: { [key in string]: string } = {
     "한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램",
 };
 
-const tempModalData: { [key in string]: string } = {
-  cautionContent:
-    "한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램",
-  interactionContent:
-    "한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램한미아스피린장용정 100밀리그램",
-};
+
 
 function FilteringSearch() {
   const [modalOpen, setModalOpen] = useState(false);
+  const handleModalOpen = (open: boolean) => {
+    setModalOpen(open)
+  }
   return (
     <>
       <FilteringSearchContainer
@@ -171,41 +170,7 @@ function FilteringSearch() {
           </BottomBox>
         </Bottom>
       </FilteringSearchContainer>
-      <Modal open={modalOpen} onClose={() => setModalOpen(false)}>
-        <ModalInnerContainer>
-          <ModalTopContainer>
-            <ModalTitleWrapper>
-              <ModalTitle $color={SUB_COLOR}>복용시 주의사항</ModalTitle>
-            </ModalTitleWrapper>
-            <ModalSubTitleWrapper>
-              <ModalSubTitle>
-                반드시 약사와의 상담 후 복용해주세요!
-              </ModalSubTitle>
-            </ModalSubTitleWrapper>
-          </ModalTopContainer>
-          <ModalCautionContainer>
-            <ModalCaution $bgColor={MAIN_COLOR}>주의사항</ModalCaution>
-            <ModalCautionContent $borderColor={MAIN_COLOR}>
-              <ModalContentScroll $scrollColor={MAIN_COLOR}>
-                {tempModalData.cautionContent}
-              </ModalContentScroll>
-            </ModalCautionContent>
-          </ModalCautionContainer>
-          <ModalInteractionContainer>
-            <ModalInteraction $bgColor={MAIN_COLOR}>상호작용</ModalInteraction>
-            <ModalInteractionContent $borderColor={MAIN_COLOR}>
-              <ModalContentScroll $scrollColor={MAIN_COLOR}>
-                {tempModalData.interactionContent}
-              </ModalContentScroll>
-            </ModalInteractionContent>
-          </ModalInteractionContainer>
-          <ModalBottom>
-            <CloseModalBtn onClick={() => setModalOpen(false)}>
-              닫기
-            </CloseModalBtn>
-          </ModalBottom>
-        </ModalInnerContainer>
-      </Modal>
+      <ResultModal modalOpen={modalOpen} handleModalOpen={handleModalOpen} />
     </>
   );
 }
